refactor: move general methods to common/schemafunc
