+++ internal/service/advancedclustertpf/plan_modifier.go
@@ -2,85 +2,16 @@ package advancedclustertpf
 
 import (
 	"context"
-	"fmt"
-	"reflect"
 	"slices"
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
-	"github.com/hashicorp/terraform-plugin-log/tflog"
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/schemafunc"
 )
 
-func IsUnknown(obj reflect.Value) bool {
-	method := obj.MethodByName("IsUnknown")
-	if !method.IsValid() {
-		panic(fmt.Sprintf("IsUnknown method not found for %v", obj))
-	}
-	results := method.Call([]reflect.Value{})
-	if len(results) != 1 {
-		panic(fmt.Sprintf("IsUnknown method must return a single value, got %v", results))
-	}
-	result := results[0]
-	response, ok := result.Interface().(bool)
-	if !ok {
-		panic(fmt.Sprintf("IsUnknown method must return a bool, got %v", result))
-	}
-	return response
-}
-
-func HasUnknowns(obj any) bool {
-	valObj := reflect.ValueOf(obj)
-	if valObj.Kind() != reflect.Ptr {
-		panic("params must be pointer")
-	}
-	valObj = valObj.Elem()
-	if valObj.Kind() != reflect.Struct {
-		panic("params must be pointer to struct")
-	}
-	typeObj := valObj.Type()
-	for i := range typeObj.NumField() {
-		field := valObj.Field(i)
-		if IsUnknown(field) {
-			return true
-		}
-	}
-	return false
-}
-
-// CopyUnknowns use reflection to copy unknown fields from src to dest.
-// The alternative without reflection would need to pass every field in a struct.
-// The implementation is similar to internal/common/conversion/model_generation.go#CopyModel
-func CopyUnknowns(ctx context.Context, src, dest any, keepUnknown []string) {
-	valSrc := reflect.ValueOf(src)
-	valDest := reflect.ValueOf(dest)
-	if valSrc.Kind() != reflect.Ptr || valDest.Kind() != reflect.Ptr {
-		panic(fmt.Sprintf("params must be pointers %v %v\n", src, dest))
-	}
-	valSrc = valSrc.Elem()
-	valDest = valDest.Elem()
-	if valSrc.Kind() != reflect.Struct || valDest.Kind() != reflect.Struct {
-		panic(fmt.Sprintf("params must be pointers to structs: %v %v\n", src, dest))
-	}
-	typeSrc := valSrc.Type()
-	typeDest := valDest.Type()
-	for i := range typeDest.NumField() {
-		fieldDest := typeDest.Field(i)
-		name := fieldDest.Name
-		if slices.Contains(keepUnknown, name) {
-			continue
-		}
-		_, found := typeSrc.FieldByName(name)
-		if !found || !IsUnknown(valDest.Field(i)) || !valDest.Field(i).CanSet() {
-			continue
-		}
-		tflog.Info(ctx, fmt.Sprintf("Copying unknown field: %s", name))
-		valDest.Field(i).Set(valSrc.FieldByName(name))
-	}
-}
-
 func useStateForUnknown(ctx context.Context, diags *diag.Diagnostics, plan, state *TFModel, keepUnknown []string) {
-	CopyUnknowns(ctx, state, plan, keepUnknown)
+	schemafunc.CopyUnknowns(ctx, state, plan, keepUnknown)
 	if slices.Contains(keepUnknown, replicationSpecsTFModelName) { // Early return if replication_specs is in keepUnknown
 		return
 	}
@@ -93,7 +24,7 @@ func useStateForUnknown(ctx context.Context, diags *diag.Diagnostics, plan, stat
 	for i := range planReplicationSpecs {
 		replicationSpecState := &stateReplicationSpecs[i]
 		replicationSpecPlan := &planReplicationSpecs[i]
-		CopyUnknowns(ctx, replicationSpecState, replicationSpecPlan, keepUnknown)
+		schemafunc.CopyUnknowns(ctx, replicationSpecState, replicationSpecPlan, keepUnknown)
 		fillInUnknownsInRegionConfigs(ctx, replicationSpecState, replicationSpecPlan, diags)
 		if diags.HasError() {
 			return
@@ -147,7 +78,7 @@ func useStateForUnknownRegionConfig(ctx context.Context, plan, state *TFRegionCo
 		}
 		plan.AutoScaling = newAutoScaling
 	}
-	CopyUnknowns(ctx, state, plan, []string{})
+	schemafunc.CopyUnknowns(ctx, state, plan, []string{})
 }
 
 func copyUnknownsFromObject[T any](ctx context.Context, diags *diag.Diagnostics, state, plan types.Object, structToObject func(context.Context, T) (basetypes.ObjectValue, diag.Diagnostics)) basetypes.ObjectValue {
@@ -156,7 +87,7 @@ func copyUnknownsFromObject[T any](ctx context.Context, diags *diag.Diagnostics,
 	if diags.HasError() {
 		return basetypes.ObjectValue{}
 	}
-	CopyUnknowns(ctx, &stateElectableSpecs, &planElectableSpecs, []string{})
+	schemafunc.CopyUnknowns(ctx, &stateElectableSpecs, &planElectableSpecs, []string{})
 	newElectableSpecs, localDiags := structToObject(ctx, planElectableSpecs)
 	if localDiags.HasError() {
 		diags.Append(localDiags...)
